Compilation and findbugs fix

+++ dcc-download-server/src/main/java/org/icgc/dcc/download/server/utils/Releases.java
@@ -60,7 +60,7 @@ public final class Releases {
     val releaseNumber = Ints.tryParse(stringReleaseNumber);
     checkArgument(releaseNumber != null, "Failed to extract release number from '%s'", releaseName);
 
-    return releaseNumber;
+    return releaseNumber.intValue();
   }
 
   public static boolean isLegacyReleaseName(@NonNull String releaseName) {
